Switch clocks to u64 instead of usize.

+++ sm83-interp/src/addressable.rs
@@ -1,4 +1,4 @@
 pub trait Addressable {
     fn read_byte(&self, index: u16) -> u8;
-    fn write_byte(&mut self, index: u16, value: u8, cpu_clock: usize);
+    fn write_byte(&mut self, index: u16, value: u8, cpu_clock: u64);
 }
